feat(watchers): gate ETH watcher code and tests behind feature flags
ETH watchers are unstable and not completed yet. This change gates all
ETH watcher functionality behind feature flags (disabled by default):

Implementation code (coins crate):
- Add `enable-eth-watchers` feature to coins/Cargo.toml
- Gate `impl WatcherOps for EthCoin` in eth.rs behind this feature
- Gate helper functions watcher_spends_hash_time_locked_payment and
  watcher_refunds_hash_time_locked_payment
- When disabled, EthCoin uses default WatcherOps impl (returns errors)

Test code (mm2_main crate):
- Convert swap_watcher_tests.rs to directory module:
  - mod.rs: shared helpers (enable_coin, enable_eth, BalanceResult, etc.)
  - utxo.rs: stable UTXO-only watcher tests (always compiled)
  - eth.rs: unstable ETH/ERC20 tests (gated behind docker-tests-watchers-eth)
- docker-tests-watchers-eth feature enables both tests and impl code

CI updates:
- Remove Geth from docker-tests-watchers job (UTXO-only now)
- Reduce container wait time from 25s to 15s
- Add _KDF_NO_GETH_DOCKER env var
