+++ mm2src/coins/Cargo.toml
@@ -10,6 +10,10 @@ run-docker-tests = []
 for-tests = ["dep:mocktopus"]
 new-db-arch = ["mm2_core/new-db-arch"]
 
+# ETH/ERC20 watcher support - disabled by default because ETH watchers are
+# unstable and not completed yet
+enable-eth-watchers = []
+
 # Temporary feature for implementing IBC wrap/unwrap mechanism and will be removed
 # once we consider it as stable.
 ibc-routing-for-swaps = []
